Update PHPMailer to v5.1 and ADOdb to v5.10
Applied db5eb5ef832329052b0c0ced2f9ad30632bb7c31 on top of ADOdb 5.10 to
patch it for PHP 5.3 compatibility reasons.

Also cleaned up non-necessary files from the library/ directory.

With respect to readme_mantis.txt files, it is easy enough to compare
the MantisBT version of a library and upstream's library to find
differences and files that have been cleaned out. No need to duplicate
information and create more work for ourselves.

There is little value to be gained from adding blank index.html files to
hide directory contents. Everyone knows the contents of the directory
already by looking at the MantisBT repository.

+++ library/adodb/adodb.inc.php
@@ -12,9 +12,9 @@
  */
 
 /**
-	\mainpage 	
+	\mainpage
 	
-	 @version V5.09 25 June 2009   (c) 2000-2009 John Lim (jlim#natsoft.com). All rights reserved.
+	 @version V5.10 10 Nov 2009   (c) 2000-2009 John Lim (jlim#natsoft.com). All rights reserved.
 
 	Released under both BSD license and Lesser GPL library license. You can choose which license
 	you prefer.
@@ -177,7 +177,7 @@
 		/**
 		 * ADODB version as a string.
 		 */
-		$ADODB_vers = 'V5.09 25 June 2009  (c) 2000-2009 John Lim (jlim#natsoft.com). All rights reserved. Released BSD & LGPL.';
+		$ADODB_vers = 'V5.10 10 Nov 2009  (c) 2000-2009 John Lim (jlim#natsoft.com). All rights reserved. Released BSD & LGPL.';
 	
 		/**
 		 * Determines whether recordset->RecordCount() is used. 
@@ -516,9 +516,6 @@
 		
 		$this->_isPersistentConnection = false;	
 			
-		global $ADODB_CACHE;
-		if (empty($ADODB_CACHE)) $this->_CreateCache();
-		
 		if ($forceNew) {
 			if ($rez=$this->_nconnect($this->host, $this->user, $this->password, $this->database)) return true;
 		} else {
@@ -584,9 +581,6 @@
 		if ($argDatabaseName != "") $this->database = $argDatabaseName;		
 			
 		$this->_isPersistentConnection = true;	
-		
-		global $ADODB_CACHE;
-		if (empty($ADODB_CACHE)) $this->_CreateCache();
 		
 		if ($rez = $this->_pconnect($this->host, $this->user, $this->password, $this->database)) return true;
 		if (isset($rez)) {
@@ -722,7 +716,7 @@
 	*  @param $table	name of table to lock
 	*  @param $where	where clause to use, eg: "WHERE row=12". If left empty, will escalate to table lock
 	*/
-	function RowLock($table,$where)
+	function RowLock($table,$where,$col='1 as ignore')
 	{
 		return false;
 	}
@@ -1307,8 +1301,8 @@
 		$ADODB_COUNTRECS = false;
 			
 
-			if ($secs2cache != 0) $rs = $this->CacheExecute($secs2cache,$sql,$inputarr);
-			else $rs = $this->Execute($sql,$inputarr);
+		if ($secs2cache != 0) $rs = $this->CacheExecute($secs2cache,$sql,$inputarr);
+		else $rs = $this->Execute($sql,$inputarr);
 		
 		$ADODB_COUNTRECS = $savec;
 		if ($rs && !$rs->EOF) {
@@ -1710,15 +1704,17 @@
 	function CacheFlush($sql=false,$inputarr=false)
 	{
 	global $ADODB_CACHE_DIR, $ADODB_CACHE;
-            
+		
+		if (empty($ADODB_CACHE)) return false;
+		
 		if (!$sql) {
 			 $ADODB_CACHE->flushall($this->debug);
-         return;
-      } 
-      
-      $f = $this->_gencachename($sql.serialize($inputarr),false);
+	         return;
+	    }
+		
+		$f = $this->_gencachename($sql.serialize($inputarr),false);
 		return $ADODB_CACHE->flushcache($f, $this->debug);
-   }
+	}
    
 	
 	/**
@@ -1767,6 +1763,8 @@
 	{
 	global $ADODB_CACHE;
 	
+		if (empty($ADODB_CACHE)) $this->_CreateCache();
+		
 		if (!is_numeric($secs2cache)) {
 			$inputarr = $sql;
 			$sql = $secs2cache;
@@ -1780,7 +1778,7 @@
 			$sqlparam = $sql;
 			
 		
-			$md5file = $this->_gencachename($sql.serialize($inputarr),true);
+		$md5file = $this->_gencachename($sql.serialize($inputarr),true);
 		$err = '';
 		
 		if ($secs2cache > 0){
@@ -1879,6 +1877,7 @@
 		$rs = $this->SelectLimit($sql,1);
 		if (!$rs) return $false; // table does not exist
 		$rs->tableName = $table;
+		$rs->sql = $sql;
 		
 		switch((string) $mode) {
 		case 'UPDATE':
@@ -2439,7 +2438,7 @@ http://www.stanford.edu/dept/itss/docs/oracle/10g/server.101/b10759/statements_1
 	{
 		if (empty($d) && $d !== 0) return 'null';
 		if ($isfld) return $d;
-
+		
 		if (is_object($d)) return $d->format($this->fmtDate);
 		
 		
@@ -2481,7 +2480,7 @@ http://www.stanford.edu/dept/itss/docs/oracle/10g/server.101/b10759/statements_1
 		if (empty($ts) && $ts !== 0) return 'null';
 		if ($isfld) return $ts;
 		if (is_object($ts)) return $ts->format($this->fmtTimeStamp);
-
+		
 		# strlen(14) allows YYYYMMDDHHMMSS format
 		if (!is_string($ts) || (is_numeric($ts) && strlen($ts)<14)) 
 			return adodb_date($this->fmtTimeStamp,$ts);
@@ -2538,8 +2537,8 @@ http://www.stanford.edu/dept/itss/docs/oracle/10g/server.101/b10759/statements_1
 		if ($rr[1] <= TIMESTAMP_FIRST_YEAR && $rr[2]<= 1) return 0;
 	
 		// h-m-s-MM-DD-YY
-		if (!isset($rr[5])) return  adodb_mktime(0,0,0,(int)$rr[2],(int)$rr[3],(int)$rr[1]);
-		return  @adodb_mktime((int)$rr[5],(int)$rr[6],(int)$rr[7],(int)$rr[2],(int)$rr[3],(int)$rr[1]);
+		if (!isset($rr[5])) return  adodb_mktime(0,0,0,$rr[2],$rr[3],$rr[1]);
+		return  @adodb_mktime($rr[5],$rr[6],$rr[7],$rr[2],$rr[3],$rr[1]);
 	}
 	
 	/**
@@ -4167,7 +4166,7 @@ http://www.stanford.edu/dept/itss/docs/oracle/10g/server.101/b10759/statements_1
 					if ($sch[1] == 'sqlite')
 						$dsna['host'] = rawurlencode($sch[1].':'.rawurldecode($dsna['host']));
 					else
-					$dsna['host'] = rawurlencode($sch[1].':host='.rawurldecode($dsna['host']));
+						$dsna['host'] = rawurlencode($sch[1].':host='.rawurldecode($dsna['host']));
 					$dsna['scheme'] = 'pdo';
 				}
 			}
